+++ src/cmdargs.cpp
@@ -1,6 +1,6 @@
 // cmdargs.cpp
 
-#include "cmdargs.h"
+#include "cmdargs.hpp"
 
 #include <iostream>
